refactor: NavBar를 Organisms로 분류
